# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ventrix API",
    "version": "1.0.0",
    "description": "API de Ventrix POS y Ventrix Contable. Generada automáticamente desde src/routes (`npm run docs:openapi`), así que refleja siempre las rutas, roles y validaciones que hay en el código.\n\nLa sesión usa un access token corto (cabecera `Authorization: Bearer`) y una cookie httpOnly de refresh. Todo es multi-tenant: el negocio sale del token, nunca del cuerpo de la petición."
  },
  "servers": [
    {
      "url": "https://api.ventrix.lat/api/v1",
      "description": "Producción"
    },
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token de un usuario (POS o Contable)."
      },
      "sellerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Sesión del portal de vendedoras (/seller), separada de la de usuarios."
      }
    },
    "schemas": {
      "RespuestaOk": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "description": "Contenido de la respuesta"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Mensaje del error"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/business/branches": {
      "get": {
        "tags": [
          "business"
        ],
        "summary": "GET /api/v1/business/branches",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "business"
        ],
        "summary": "Solo el dueño del negocio (ADMIN) puede crear/editar bodegas — nunca el staff. El límite de plan (planLimit.branches()) es el único freno duro: Free se queda en 1, Pro en 2 (ver plans.ts) — a propósito no ilimitado, para que no se use una sola suscripción Pro para operar más de dos locales.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/business/branches/{id}": {
      "put": {
        "tags": [
          "business"
        ],
        "summary": "PUT /api/v1/business/branches/{id}",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/business/me": {
      "get": {
        "tags": [
          "business"
        ],
        "summary": "GET /api/v1/business/me",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "business"
        ],
        "summary": "PUT /api/v1/business/me",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email inválido"
                  },
                  "taxRate": {
                    "type": "number",
                    "description": "IVA debe estar entre 0 y 100",
                    "minimum": 0,
                    "maximum": 100
                  },
                  "name": {
                    "type": "string",
                    "description": "El nombre no puede estar vacío"
                  },
                  "currency": {
                    "type": "string",
                    "description": "Moneda debe ser un código de 3 letras",
                    "minLength": 3,
                    "maxLength": 3
                  },
                  "settings": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/business/payment-accounts": {
      "get": {
        "tags": [
          "business"
        ],
        "summary": "GET /api/v1/business/payment-accounts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "business"
        ],
        "summary": "POST /api/v1/business/payment-accounts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "El nombre es requerido"
                  },
                  "type": {
                    "type": "string",
                    "description": "Tipo inválido"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        }
      }
    },
    "/api/v1/business/payment-accounts/{id}": {
      "delete": {
        "tags": [
          "business"
        ],
        "summary": "Borrado seguro: si el medio ya tiene movimientos (ventas/compras/gastos/abonos), NO se borra — se desactiva, para conservar a qué cuenta fue cada movimiento histórico. Solo se elimina de verdad si nunca se usó.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "put": {
        "tags": [
          "business"
        ],
        "summary": "PUT /api/v1/business/payment-accounts/{id}",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "El nombre no puede estar vacío"
                  },
                  "type": {
                    "type": "string",
                    "description": "Tipo inválido"
                  },
                  "active": {
                    "type": "boolean"
                  },
                  "order": {
                    "type": "integer",
                    "minimum": 0
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}